+++ CHANGELOG.md
@@ -10,8 +10,11 @@ MolSimToolkit.jl Changelog
 [badge-fix]: https://img.shields.io/badge/Fix-purple.svg
 [badge-info]: https://img.shields.io/badge/Info-gray.svg
 
-Version 1.32.4-DEV
+Version 1.33.0-DEV
 --------------
+- ![FEATURE][badge-feature] Introduce the option to compute the RMSD of a structure, given the alignment of other structure (or parts of); i. e. the `rmsd_of` option of `rmsd`. 
+- ![FEATURE][badge-feature] Support string selections in `rmsd` and `rmsd_matrix` functions.
+- ![ENHANCEMENT][badge-enhancement] Reconstruct polymer and complex structure before computing rmsds, to avoid issues with broken molecule coordinates through PBCs.
 
 Version 1.32.3
 --------------